Fixed bug and simplified codebase
Instead of having the NavMenu interact with other components to
hide content behind it, I instead had the active menu overlay
go over whatever content is currently on the page. This helps
simplify UI for the user, keeps the already loaded resources on the
page and fixes a bug that allows users to maintain their place on the
site when the nav menu is active.

Switching how we handle the modal menu also cleaned up the codebase
and prevents us from using a hook.